refactor: remove ghost-fetch, use impit directly
Actor has been broken for months. Now it calls the Expedia mobile API directly through impit with residential proxy. Input and output unchanged.
